pyproject.toml
[project]
name = "first-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "fastapi>=0.115.8",
    "greenlet>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
]
